Create MarksPortfolio.md
My portfolio outlies my skills and ability to contribute as a software developer. These skilled coupled with my business acumen, and leadership accomplishments qualify me to be an excellent addition to your team. I have a unique perspective and ability to understand your business needs, and contribute to your success.